added vr button into develop menu added comment about the early return within the renderloop replaced standard VR Button class with custom menu item class
